major code cleanup, changed to simplified AST, distinction between resolving a value (looking up a variable) and evaluating a value. Also changed to all native types in AST, with addition of the Symbol class, allowing code-as-data and data-as-code.

+++ linkedlisp/src/main/java/com/googlecode/linkedlisp/functions/math/Plus.java
@@ -11,7 +11,7 @@ public class Plus extends ArithmeticFunction {
 		return "+";
 	}
 
-	public Float operation (Float current, Float factor) {
+	public Double operation (Double current, Double factor) {
 		current += factor;
 		return current;
 	}
